Updated deleteQueryBatch callback to be optional.
