+++ src/utils/pubsub.ts
@@ -1,7 +1,6 @@
 import {db} from "../index";
 import {deleteCollection} from "./misc";
 import {ScheduledEvent} from "firebase-functions/lib/v2/providers/scheduler";
-import {BatchUtil} from "./batch";
 
 async function trackProcessedIds(topicName: string, id: string) {
   const docRef = db.doc(`@topics/${topicName}/processedIds/${id}`);
@@ -27,14 +26,7 @@ export async function cleanPubSubProcessedIds(event: ScheduledEvent) {
     const query = topicDoc.ref.collection("processedIds")
       .where("timestamp", "<", new Date(Date.now() - 1000 * 60 * 60 * 24 * 7));
 
-    await deleteCollection(query, async (snapshot) => {
-      const batch = BatchUtil.getInstance();
-      snapshot.docs.forEach( (doc) => {
-        batch.deleteDoc(doc.ref);
-      });
-
-      await batch.commit();
-    });
+    await deleteCollection(query);
     i++;
   }
   console.info(`Cleaned ${i} topics of processedIds`);
